fix: Float not Double

+++ Sources/KDNAAppShared/Models/ReasoningEffort.swift
@@ -86,7 +86,7 @@ public enum ReasoningEffort: String, Codable, CaseIterable, Sendable {
 }
 
 public struct GenerationSettings: Codable, Sendable, Equatable {
-    public init(temperature: Double = 0.7, reasoningEffort: ReasoningEffort = .medium) {
+    public init(temperature: Float = 0.7, reasoningEffort: ReasoningEffort = .medium) {
         self.temperature = temperature; self.reasoningEffort = reasoningEffort
     }
     public var temperature: Float
